requirements.txt
colored==2.2.3
gmpy2==2.1.2
pycryptodome==3.10.4
Requests==2.31.0
